+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: yhasegaw <yhasegaw@student.42.jp>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/14 15:06:13 by yhasegaw          #+#    #+#             */
-/*   Updated: 2020/10/28 21:15:09 by yhasegaw         ###   ########.fr       */
+/*   Updated: 2020/10/30 02:19:24 by yhasegaw         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 
 	if (!s)
 		return (NULL);
-	if (len == 0 || start >= ft_strlen(s))
+	if (len == 0 || start >= (unsigned int)ft_strlen(s))
 		return (ft_strdup(""));
 	if (!(ret = (char *)malloc(sizeof(char) * (len + 1))))
 		return (NULL);
